backend/requirements.txt
# Web
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Config & Validation
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0

# Auth / Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
google-auth==2.23.0

# AI / LLM - ATUALIZADO
google-generativeai>=0.8.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0